requirements.txt
numpy==1.14.0
Pillow==5.0.0
PyYAML==3.12
six==1.11.0
torch==0.3.1
torchvision==0.2.0
